Makefile now behaves like docs say + misc - fixes #2
Added 'get' and 'format' targets to the Makefile according to the
existing documentation in the README.md, as well as 'all' as the
first target which runs both (i.e. when make is called with no args).

Commented out call to 'check-logins.coffee' since it is apparently
just a debug thing which doesn't work the first time due to an
ordering problem (needs data which doesn't get generated until
later). This, along with earlier commit which checks that
temp-logins.json exists before trying to rename it to old-logins,
fixes #2 (https://github.com/paulmillr/top-github-users/issues/2)

Now running 'make' out of the box after cloning and installing deps
should just work...

Also added a 'clean' target to the Makefile for good measure, and
just as a means to document the list of files that are expected to
be generated from get/format

Also updated the request error handling code to throw the body text
in addition to the status (useful to be able to see if the error is
because you've run afoul of Github's rate limit rules).

Finally, split url string in get-users.coffee into two lines for
readability.